fix(qtip): widen low-rate traceback offsets

+++ banana-smasher/src/banana_smasher/qtip25_native_v4.py
@@ -1050,7 +1050,7 @@ __global__ void qtip3_viterbi_kernel(
         }
         costs[step & 1][prefix] = best;
         if (step >= capture_step) {
-          choices[((step - capture_step) * batch + sequence) * PREFIXES + prefix] =
+          choices[(static_cast<int64_t>(step - capture_step) * batch + sequence) * PREFIXES + prefix] =
               best_state;
         }
       }
@@ -1096,7 +1096,7 @@ __global__ void qtip3_viterbi_kernel(
         }
         costs[step & 1][tid] = best;
         if (step >= capture_step) {
-          choices[((step - capture_step) * batch + sequence) * PREFIXES + tid] =
+          choices[(static_cast<int64_t>(step - capture_step) * batch + sequence) * PREFIXES + tid] =
               best_state;
         }
       }
@@ -1119,7 +1119,7 @@ __global__ void qtip3_viterbi_kernel(
     }
     for (int step = steps - 1; step >= capture_step; --step) {
       const int state =
-          choices[((step - capture_step) * batch + sequence) * PREFIXES + prefix];
+          choices[(static_cast<int64_t>(step - capture_step) * batch + sequence) * PREFIXES + prefix];
       if (!overlap_only) {
         output[step * batch + sequence] = state;
       }
